refactor(AudioLibraryScanner): move `AudioFilesScanner` to another file
